Restructured tha AssignedTA controller a bit
